+++ buildVars.py
@@ -24,11 +24,11 @@ addon_info = {
 	Otherwise, enter in the edit field the word you want, tab an choose the dictionary you want and press enter.
 	The meaning of the word will be displayed in a separate browseable message box."""),
 	# version
-	"addon_version": "2.6.1",
+	"addon_version": "2.6.2",
 	# Author(s)
 	"addon_author" : u"Ibrahim Hamadeh<ibra.hamadeh@hotmail.com>",
 	# URL for the add-on documentation support
-	"addon_url" : "https://github.com/ibrahim-h/dictionariesAlmaany.git",
+	"addon_url" : "https://github.com/ibrahim-h/dictionariesAlmaany",
 	# Documentation file name
 	"addon_docFileName" : "readme.html",
 	# Minimum NVDA version supported (e.g. "2018.3")